+++ app/services/role.service.js
@@ -1,14 +1,14 @@
-const roleRepository = require('../repositories/role.repositories')
+const roleRepository = require('../repositories/role.repository');
 
 const getAllRoles = async() => {
     return roleRepository.getAll();
-}
+};
 
 const getRole = async(id) => {
     return roleRepository.findById(id);
-}
+};
 
 module.exports = {
     getAllRoles,
     getRole
-}
+};
